missing-routes
add missing routes
